Crate mango_v4_client

Modules

account_update_stream

chain_data

error_tracking

gpa

health_cache

jupiter

perp_pnl

snapshot_source

websocket_source

Structs

CachedAccountFetcher

Client

ClientBuilder

Builder for Client.

ComputeEstimates

MangoClient

MangoGroupContext

PerpMarketContext

PreparedInstructions

RpcAccountFetcher

Serum3MarketContext

TokenContext

TransactionBuilder

TransactionBuilderConfig

TransactionSize

Enums

ClientBuilderError

Error type for ClientBuilder

JupiterSwapMode

MangoClientError

Constants

MAX_ACCOUNTS_PER_TRANSACTION

Traits

AccountFetcher

AnyhowWrap

Some Result<> types don’t convert to anyhow::Result nicely. Force them through stringification.

AsyncChannelSendUnlessFull

Push to an async_channel::Sender and ignore if the channel is full

Functions

account_fetcher_fetch_anchor_account

account_fetcher_fetch_mango_account

delay_interval

Like tokio::time::interval(), but with Delay as default MissedTickBehavior

http_error_handling

keypair_from_cli

prettify_client_error

Do some manual unpacking on some ClientErrors

prettify_solana_client_error

pubkey_from_cli

send_and_confirm_transaction

A copy of RpcClient::send_and_confirm_transaction that returns the slot the transaction confirmed in.

to_readonly_account_meta

to_writable_account_meta

tracing_subscriber_init

Convenience function used in binaries to set up the fmt tracing_subscriber, with cololring enabled only if logging to a terminal and with EnvFilter.

Type Definitions

SimulateTransactionResponse