Struct mango_v4_client::MangoGroupContext

pub struct MangoGroupContext {
    pub group: Pubkey,
    pub tokens: HashMap<TokenIndex, TokenContext>,
    pub token_indexes_by_name: HashMap<String, TokenIndex>,
    pub serum3_markets: HashMap<Serum3MarketIndex, Serum3MarketContext>,
    pub serum3_market_indexes_by_name: HashMap<String, Serum3MarketIndex>,
    pub perp_markets: HashMap<PerpMarketIndex, PerpMarketContext>,
    pub perp_market_indexes_by_name: HashMap<String, PerpMarketIndex>,
    pub address_lookup_tables: Vec<Pubkey>,
    pub compute_estimates: ComputeEstimates,
}

Fields
group: Pubkeytokens: HashMap<TokenIndex, TokenContext>token_indexes_by_name: HashMap<String, TokenIndex>serum3_markets: HashMap<Serum3MarketIndex, Serum3MarketContext>serum3_market_indexes_by_name: HashMap<String, Serum3MarketIndex>perp_markets: HashMap<PerpMarketIndex, PerpMarketContext>perp_market_indexes_by_name: HashMap<String, PerpMarketIndex>address_lookup_tables: Vec<Pubkey>compute_estimates: ComputeEstimates
Implementations

impl MangoGroupContext

pub fn mint_info_address(&self, token_index: TokenIndex) -> Pubkey

pub fn perp(&self, perp_market_index: PerpMarketIndex) -> &PerpMarketContext

pub fn perp_market_address(&self, perp_market_index: PerpMarketIndex) -> Pubkey

pub fn serum3_market_index(&self, name: &str) -> Serum3MarketIndex

pub fn serum3(&self, market_index: Serum3MarketIndex) -> &Serum3MarketContext

pub fn serum3_base_token(&self, market_index: Serum3MarketIndex) -> &TokenContext

pub fn serum3_quote_token( &self, market_index: Serum3MarketIndex ) -> &TokenContext

pub fn token(&self, token_index: TokenIndex) -> &TokenContext

pub fn token_by_mint(&self, mint: &Pubkey) -> Result<&TokenContext>

pub fn token_by_name(&self, name: &str) -> &TokenContext

pub async fn new_from_rpc(rpc: &RpcClientAsync, group: Pubkey) -> Result<Self>

pub fn derive_health_check_remaining_account_metas( &self, account: &MangoAccountValue, affected_tokens: Vec<TokenIndex>, writable_banks: Vec<TokenIndex>, affected_perp_markets: Vec<PerpMarketIndex>, fallback_contexts: HashMap<Pubkey, FallbackOracleContext> ) -> Result<(Vec<AccountMeta>, u32)>

pub fn derive_health_check_remaining_account_metas_two_accounts( &self, account1: &MangoAccountValue, account2: &MangoAccountValue, affected_tokens: &[TokenIndex], writable_banks: &[TokenIndex], fallback_contexts: HashMap<Pubkey, FallbackOracleContext> ) -> Result<(Vec<AccountMeta>, u32)>

pub fn changed_significantly(&self, other: &Self) -> bool

Returns true if the on-chain context changed significantly, this currently means:

new listings (token, serum, perp)

oracle pubkey or config changes

other config changes visible through the context This is done because those would affect the pubkeys the websocket streams need to listen to, or change limits, oracle staleness or other relevant configuration.

pub async fn new_tokens_listed(&self, rpc: &RpcClientAsync) -> Result<bool>

pub async fn new_serum3_markets_listed( &self, rpc: &RpcClientAsync ) -> Result<bool>

pub async fn new_perp_markets_listed(&self, rpc: &RpcClientAsync) -> Result<bool>

pub async fn derive_fallback_oracle_keys( &self, fallback_oracle_config: &FallbackOracleConfig, account_fetcher: &dyn AccountFetcher ) -> Result<HashMap<Pubkey, FallbackOracleContext>>

Returns a map of oracle pubkey -> FallbackOracleContext